cmd: sf: fix [XJ3-4176] Fix potential variable overflow from get_timer
Summary:
Since get_timer() returns ulong, fix functions using unsigned int
to store get_timer result to aviod potential variable overflow.


Test Plan: build okay, net and sf functionality okay

Reviewers: ming.yu, yan04.zhang, yunqian.wang

Reviewed By: ming.yu, yunqian.wang

Subscribers: qinlong.wang

Differential Revision: https://cr.hobot.cc/D181673

+++ cmd/sf.c
@@ -402,8 +402,8 @@ static char *stage_name[STAGE_COUNT] = {
 struct test_info {
 	int stage;
 	int bytes;
-	unsigned base_ms;
-	unsigned time_ms[STAGE_COUNT];
+	ulong base_ms;
+	ulong time_ms[STAGE_COUNT];
 };
 
 static void show_time(struct test_info *test, int stage)
